aider: Changed buffer filetype from markdown to custom filetype 'flyboy' in chat.lua
# Aider chat conversation:

USER: add a custom filetype instead of markdown to the buffer created and then adopt the markdown syntax for that custom `filetype`

ASSISTANT: To achieve this, we need to modify the `open_chat_with_text` function in the `lua/flyboy/chat.lua` file. We will change the filetype from "markdown" to a custom filetype, let's say "flyboy".

Here is the *edit block* for this change:

```lua
lua/flyboy/chat.lua
<<<<<<< HEAD
    local buffer = vim.api.nvim_create_buf(true, false)
    vim.api.nvim_buf_set_option(buffer, "filetype", "markdown")
=======
    local buffer = vim.api.nvim_create_buf(true, false)
    vim.api.nvim_buf_set_option(buffer, "filetype", "flyboy")
>>>>>>> updated
```

After this change, you need to set up the syntax highlighting for the new filetype "flyboy". This is typically done in a Vim syntax file. However, since the syntax file is not provided in the list of *read-write* files, you need to create it yourself.

The syntax file should be named `flyboy.vim` and located in the `syntax` directory of your Vim configuration directory. In this file, you can include the markdown syntax file to adopt the markdown syntax for the "flyboy" filetype. The content of the `flyboy.vim` file should be:

```vim
runtime! syntax/markdown.vim
```

Please note that this file is not included in the list of *read-write* files, so I can't provide an *edit block* for it. You need to create this file yourself.