aider: Added markdown syntax setting to chat.lua file.
# Aider chat conversation:

USER: I did `git reset --hard HEAD~1` to discard the last edits.

ASSISTANT: Okay, that means the changes you made in the last commit have been discarded and your repository is now at the state it was before that commit. If you need to make any changes or need assistance with anything else, feel free to ask.

USER: can I instead set the syntax in the chat.lua file?

ASSISTANT: Yes, you can set the syntax in the `chat.lua` file. This can be done in the `open_chat_with_text` function where the buffer is created and the filetype is set to "flyboy". After setting the filetype, you can set the syntax to "markdown".

Here is the change you need to make:

```lua
lua/flyboy/chat.lua
<<<<<<< HEAD
    local function open_chat_with_text(text)
        -- create a new empty buffer
        local buffer = vim.api.nvim_create_buf(true, false)
        vim.api.nvim_buf_set_option(buffer, "filetype", "flyboy")
        local lines = vim.split(text, "\n")
=======
    local function open_chat_with_text(text)
        -- create a new empty buffer
        local buffer = vim.api.nvim_create_buf(true, false)
        vim.api.nvim_buf_set_option(buffer, "filetype", "flyboy")
        vim.api.nvim_buf_set_option(buffer, "syntax", "markdown")
        local lines = vim.split(text, "\n")
>>>>>>> updated
```

This change will set the syntax to "markdown" for every new buffer that is created with the filetype "flyboy".